Bug Fixed: Replaced custom hook for a new context
I confused the concept of custom hooks and context. When the custom hook was used in the same component (app.js) everything worked fine, but when I used one state in one component and another state in the other component the first one wasn't updated (because it was two different instances of the hook).

So I created a context, and reused the same logic of the previous context and everything worked fine